+++ src/structs/mod.rs
@@ -1,0 +1,2 @@
+pub mod user;
+pub mod wallet;
